Refactor Actor.Performance
I moved it out into their own files.
Non-actor based events can also be handled with these scripts.
I haven't gotten there yet.

I decided to use GDScript, because using the dynamically
typed dictionaries in C# was a paiiinnn.

I was thinking of remaking literally the entire view in GDScript,
but that is too dramatic of a change. Thinking about it was paralyzing.
It still might be a good idea due to shenanigans with static typing.

Also, I might remove the Godot from the Model.
I only really was using it for composing objects with scenes,
but that can be done outside of Godot. It makes the game crash on
hot reload, due to using tuples everywhere.
(man, i want Vector2i from the beta.)