Distances measured in Tile units
I noticed the map vision behaving poorly.
It assumed radius was in tiles, but it was half-tiles.
This broke, so I fixed it naively.
This lead to using 4x as large an array as needed.

Fixing this should help against future mixups.
It is annoying to type the .5, but thats a small price.