random changes. see description :P
It's been a week; I hit a slump. Anyways, I got advice and more feedback.
This might be weirdly worded, but its late and I just want to get working.

The game is blowing up in scope. There are problems I cannot tackle
on my own, like AI, random generation, and just content.
Art, audio, and UI are important, but I'm not interested in
them (at the moment). This makes the game confusing as it doesn't
give the player the required info. I'll have to simplify or get help
to get the game "done" in a timely manner.

Also, the combat is not functioning well. I'll need to experiment,
and either build upon it or toss it out.
(Very few people paid attention to the crit => stun thing, and
just spammed the attack button while waiting for animations to finish.
I'm considering also dropping animations for the same reason.)

The immediate goal is a vertical slice. That entails
enough skills for meaningful combat/engagements, a better
victory condition/points system, a more informational UI,
and some "balancing" of risk management?

This is in contrast to the mashy combat, the lack of extrinsic
motivation to explore further, the "hidden" energy stat
and weird turn taking, and the unmitigatable risk in combat.