Fix hotpatching.
Godot doesn't have trouble with tuples, it just ignores them
when trying to serialize. On reload, itll reinitialize to a default value.
(EX position in Entity. All entities move to (0, 0) on reload.)

Godot does have trouble with dictionaries with tuple values (and keys?).
It can't ignore the dictionary entry, so it hard crashes.

This could cause glitchiness on hotpatching. Lots of values could be
reinitialized. I'll have to be careful with that.