suppose it works.
"it" being removing Wait and SmallWait events.
not feeling confident about it.
maybe because its 3 am.

the logic is awkward, and it isn't leveraging the engine much.
also don't know how powerful the new approach is.
(it only has access to the current and previous event.)

i was thinking of just dropping animations and using "impatient mode"
to develop. make it more like an ascii game in the terminal.