Rework Handler Flow. Abandon Experiment.
Feels really good to make big moves, but I doubt this is the one.

The original intent is to group events together into larger units.
This way, when handling events, an entire unit will be finished before
moving onto the next.

However, the solution is to keep a list of active or running events,
and only run new events if they're compatible.
This is not a new idea. I had been too clever and tried to never wait
for events to finish and only wait for animations for timing.

I can still cherrypick some commits to go to development,
so its not a complete bust.