Namespace Engine Helpers.
Large (existing) API surface area.
If I'm bored I can trim it down and hide these in the namespace.

For some of these, it makes sense to keep them in the Engine directory.
Vector2i and AbsolutePosition are params and return values.

GridHelper and VisibilityTrie could be rolled into Model, but ehh.

SparseMatrix could definitely be rolled into Model or some interface.